Implement getElementsByClassName using recursion
This function returns an array of elements which contain the class name input into the function.

Using element.childNodes, which gives an array of childeren, we can chech if the class name is contained in the element class list and if it is we push the current element to the array. If the element contains children we loop through each child and check if the class name is contained within its class list and push it to the array if it is. We also check if there are any childrent of the children array and so on and so forth. document.body is the main parent of the page where we start the recursive search function.